move gbfiletype to utils; switch to test acc that is not problematic
